added smoke test for logit lens
