+++ tuned_lens/nn/__init__.py
@@ -4,4 +4,4 @@ from .decoder import (
 )
 from .downstream_wrapper import DownstreamWrapper
 from .probe_dict import ProbeDict
-from .tuned_lens import TunedLens
+from .lenses import TunedLens
