+++ tuned_lens/nn/downstream_wrapper.py
@@ -1,4 +1,4 @@
-from .tuned_lens import TunedLens
+from .lenses import TunedLens
 from ..utils import pytree_map
 from transformers import PreTrainedModel, PreTrainedTokenizerBase
 from typing import Iterable, NamedTuple, Optional, Sequence
